feat(server.ts): add support for AWS Lambda integration using @as-integrations/aws-lambda package
The changes in this commit add support for integrating the server with AWS Lambda using the `@as-integrations/aws-lambda` package. This allows the server to be deployed as a Lambda function and handle API Gateway proxy events.

The `startServerAndCreateLambdaHandler` function from the package is used to start the Apollo Server and create a Lambda handler function. The `graphqlHandler` export is then used as the Lambda handler.

Additionally, the `handlers.createAPIGatewayProxyEventV2RequestHandler()` function is passed as an argument to the `startServerAndCreateLambdaHandler` function to use the Proxy V2 handler for API Gateway events.