feat(docker-compose.yml): add docker-compose file for running a local DynamoDB instance
A new docker-compose.yml file has been added to the backend/db/event-store directory. This file sets up a local DynamoDB instance using the amazon/dynamodb-local image. The container is named dynamodb-local and is configured to expose port 8000. The data directory for the DynamoDB instance is mapped to ./docker/dynamodb on the host machine.