feat(workspace.ts): add equals method to Workspace class for comparing workspace objects
The equals method is added to the Workspace class to enable comparison between two Workspace objects. It checks if the id, name, sequenceNumber, and version of the current Workspace object are equal to the corresponding properties of the other Workspace object. This method can be used to determine if two Workspace objects are equal.